2022-9-20 上线脚本微调

+++ 上线脚本/舆情风险信息(ok)/DDL/舆情风险信息整合DDL.sql
@@ -18,7 +18,7 @@ create table pth_rmp.rmp_opinion_risk_info
 	url_kw STRING,
 	news_from STRING,
 	msg STRING,
-	delete_flag TINYINT,
+	delete_flag int,
 	create_by STRING,
 	create_time TIMESTAMP,
 	update_by STRING,
